+++ lib/views/screens/home/home_screen.dart
@@ -1,7 +1,13 @@
 import 'package:expense_tracker/data/models/expense_model/expense_model.dart';
 import 'package:expense_tracker/res/colors/app_colors.dart';
+import 'package:expense_tracker/res/images/app_images.dart';
 import 'package:expense_tracker/res/text_styles/app_texts.dart';
 import 'package:expense_tracker/viewmodels/home/home_viewmodel.dart';
+import 'package:expense_tracker/views/screens/allTransactions/all_transactions.dart';
+import 'package:expense_tracker/views/screens/bottomNavBar/bottomNavBar.dart';
+import 'package:expense_tracker/views/screens/home/widgets/accountBalanceCard.dart';
+import 'package:expense_tracker/views/screens/home/widgets/timeFilter.dart';
+import 'package:expense_tracker/views/screens/home/widgets/transactionList.dart';
 import 'package:expense_tracker/views/widgets/transaction_card.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
@@ -45,13 +51,24 @@ class _HomeScreenState extends State<HomeScreen> {
                     Row(
                       mainAxisAlignment: MainAxisAlignment.spaceBetween,
                       children: [
-                        const CircleAvatar(
-                          radius: 20,
-                          backgroundImage: AssetImage(
-                            'assets/images/profile.png',
+                        GestureDetector(
+                          onTap: () {
+                            Navigator.push(
+                              context,
+                              MaterialPageRoute(
+                                builder: (_) => const ProfileScreen(),
+                              ),
+                            );
+                          },
+                          child: CircleAvatar(
+                            radius: 20,
+                            backgroundImage: const AssetImage(
+                              AppImages.userIcon,
+                            ),
                           ),
                         ),
-                        Text('Cipher X', style: AppTextStyles.subheading),
+                        Text('Expense Tracker',
+                            style: AppTextStyles.subheading),
                         IconButton(
                           icon: const Icon(
                             Icons.notifications_outlined,
@@ -64,188 +81,12 @@ class _HomeScreenState extends State<HomeScreen> {
                     const SizedBox(height: 24),
 
                     // Account balance card
-                    Container(
-                      padding: const EdgeInsets.all(16),
-                      decoration: BoxDecoration(
-                        color: AppColors.cardColor,
-                        borderRadius: BorderRadius.circular(16),
-                        boxShadow: [
-                          BoxShadow(
-                            color: Colors.black.withOpacity(0.05),
-                            blurRadius: 10,
-                            offset: const Offset(0, 5),
-                          ),
-                        ],
-                      ),
-                      child: Column(
-                        crossAxisAlignment: CrossAxisAlignment.start,
-                        children: [
-                          const Text(
-                            'Account Balance',
-                            style: AppTextStyles.caption,
-                          ),
-                          const SizedBox(height: 8),
-                          Text(
-                            '₹${homeViewModel.balance.toStringAsFixed(0)}',
-                            style: AppTextStyles.heading.copyWith(
-                              fontSize: 32,
-                              fontWeight: FontWeight.w700,
-                            ),
-                          ),
-                          const SizedBox(height: 16),
-                          Row(
-                            children: [
-                              // Income box
-                              Expanded(
-                                child: Container(
-                                  padding: const EdgeInsets.all(12),
-                                  decoration: BoxDecoration(
-                                    color: AppColors.incomeColor.withOpacity(
-                                      0.1,
-                                    ),
-                                    borderRadius: BorderRadius.circular(12),
-                                  ),
-                                  child: Column(
-                                    crossAxisAlignment:
-                                        CrossAxisAlignment.start,
-                                    children: [
-                                      Row(
-                                        children: [
-                                          Container(
-                                            padding: const EdgeInsets.all(6),
-                                            decoration: BoxDecoration(
-                                              color: AppColors.incomeColor,
-                                              borderRadius:
-                                                  BorderRadius.circular(8),
-                                            ),
-                                            child: const Icon(
-                                              Icons.arrow_downward,
-                                              color: Colors.white,
-                                              size: 16,
-                                            ),
-                                          ),
-                                          const SizedBox(width: 8),
-                                          const Text(
-                                            'Income',
-                                            style: AppTextStyles.caption,
-                                          ),
-                                        ],
-                                      ),
-                                      const SizedBox(height: 8),
-                                      Text(
-                                        '₹${homeViewModel.totalIncomes.toStringAsFixed(0)}',
-                                        style: AppTextStyles.subheading,
-                                      ),
-                                    ],
-                                  ),
-                                ),
-                              ),
-                              const SizedBox(width: 16),
-                              // Expense box
-                              Expanded(
-                                child: Container(
-                                  padding: const EdgeInsets.all(12),
-                                  decoration: BoxDecoration(
-                                    color: AppColors.expenseColor.withOpacity(
-                                      0.1,
-                                    ),
-                                    borderRadius: BorderRadius.circular(12),
-                                  ),
-                                  child: Column(
-                                    crossAxisAlignment:
-                                        CrossAxisAlignment.start,
-                                    children: [
-                                      Row(
-                                        children: [
-                                          Container(
-                                            padding: const EdgeInsets.all(6),
-                                            decoration: BoxDecoration(
-                                              color: AppColors.expenseColor,
-                                              borderRadius:
-                                                  BorderRadius.circular(8),
-                                            ),
-                                            child: const Icon(
-                                              Icons.arrow_upward,
-                                              color: Colors.white,
-                                              size: 16,
-                                            ),
-                                          ),
-                                          const SizedBox(width: 8),
-                                          const Text(
-                                            'Expenses',
-                                            style: AppTextStyles.caption,
-                                          ),
-                                        ],
-                                      ),
-                                      const SizedBox(height: 8),
-                                      Text(
-                                        '₹${homeViewModel.totalExpenses.toStringAsFixed(0)}',
-                                        style: AppTextStyles.subheading,
-                                      ),
-                                    ],
-                                  ),
-                                ),
-                              ),
-                            ],
-                          ),
-                        ],
-                      ),
-                    ),
+                    accountBalanceCard(homeViewModel: homeViewModel),
 
                     const SizedBox(height: 24),
 
                     // Time period filter
-                    Container(
-                      height: 40,
-                      decoration: BoxDecoration(
-                        color: Colors.grey.withOpacity(0.1),
-                        borderRadius: BorderRadius.circular(20),
-                      ),
-                      child: Row(
-                        children: [
-                          Expanded(
-                            child: Container(
-                              decoration: BoxDecoration(
-                                color: AppColors.primary.withOpacity(0.1),
-                                borderRadius: BorderRadius.circular(20),
-                              ),
-                              alignment: Alignment.center,
-                              child: Text(
-                                'Today',
-                                style: AppTextStyles.caption.copyWith(
-                                  color: AppColors.primary,
-                                  fontWeight: FontWeight.w600,
-                                ),
-                              ),
-                            ),
-                          ),
-                          const Expanded(
-                            child: Center(
-                              child: Text(
-                                'Week',
-                                style: AppTextStyles.caption,
-                              ),
-                            ),
-                          ),
-                          const Expanded(
-                            child: Center(
-                              child: Text(
-                                'Month',
-                                style: AppTextStyles.caption,
-                              ),
-                            ),
-                          ),
-                          const Expanded(
-                            child: Center(
-                              child: Text(
-                                'Year',
-                                style: AppTextStyles.caption,
-                              ),
-                            ),
-                          ),
-                        ],
-                      ),
-                    ),
+                    TimeFilter(),
 
                     const SizedBox(height: 24),
 
@@ -258,7 +99,14 @@ class _HomeScreenState extends State<HomeScreen> {
                           style: AppTextStyles.subheading,
                         ),
                         TextButton(
-                          onPressed: () {},
+                          onPressed: () {
+                            Navigator.push(
+                              context,
+                              MaterialPageRoute(
+                                builder: (_) => const AllTransactionsScreen(),
+                              ),
+                            );
+                          },
                           child: Text(
                             'See All',
                             style: AppTextStyles.caption.copyWith(
@@ -272,49 +120,7 @@ class _HomeScreenState extends State<HomeScreen> {
                     const SizedBox(height: 8),
 
                     // Transaction list
-                    ListView.separated(
-                      shrinkWrap: true,
-                      physics: const NeverScrollableScrollPhysics(),
-                      itemCount: homeViewModel.recentTransactions.length > 5
-                          ? 5
-                          : homeViewModel.recentTransactions.length,
-                      separatorBuilder: (context, index) =>
-                          const SizedBox(height: 8),
-                      itemBuilder: (context, index) {
-                        final transaction =
-                            homeViewModel.recentTransactions[index];
-                        final bool isExpense = transaction is Expense;
-
-                        return Dismissible(
-                          key: Key(transaction.id),
-                          direction: DismissDirection.endToStart,
-                          background: Container(
-                            padding: const EdgeInsets.only(right: 20),
-                            alignment: Alignment.centerRight,
-                            decoration: BoxDecoration(
-                              color: Colors.red,
-                              borderRadius: BorderRadius.circular(12),
-                            ),
-                            child: const Icon(
-                              Icons.delete,
-                              color: Colors.white,
-                            ),
-                          ),
-                          onDismissed: (direction) {
-                            // Handle delete action
-                            Provider.of<HomeViewModel>(context, listen: false)
-                                .deleteTransaction(transaction);
-                          },
-                          child: TransactionCard(
-                            category: transaction.category,
-                            description: transaction.description,
-                            amount: transaction.amount,
-                            isExpense: isExpense,
-                            date: transaction.date,
-                          ),
-                        );
-                      },
-                    ),
+                    TransactionList(homeViewModel: homeViewModel),
                   ],
                 ),
               ),
@@ -379,66 +185,13 @@ class _HomeScreenState extends State<HomeScreen> {
         child: const Icon(Icons.add, color: Colors.white),
       ),
       floatingActionButtonLocation: FloatingActionButtonLocation.centerDocked,
-      bottomNavigationBar: BottomAppBar(
-        shape: const CircularNotchedRectangle(),
-        notchMargin: 8,
-        child: SizedBox(
-          height: 60,
-          child: Row(
-            mainAxisAlignment: MainAxisAlignment.spaceAround,
-            children: [
-              IconButton(
-                icon: Icon(
-                  Icons.home,
-                  color: _currentIndex == 0 ? AppColors.primary : Colors.grey,
-                ),
-                onPressed: () {
-                  setState(() {
-                    _currentIndex = 0;
-                  });
-                },
-              ),
-              IconButton(
-                icon: Icon(
-                  Icons.bar_chart,
-                  color: _currentIndex == 1 ? AppColors.primary : Colors.grey,
-                ),
-                onPressed: () {
-                  setState(() {
-                    _currentIndex = 1;
-                  });
-                },
-              ),
-              const SizedBox(width: 40), // For FAB
-              IconButton(
-                icon: Icon(
-                  Icons.notifications,
-                  color: _currentIndex == 2 ? AppColors.primary : Colors.grey,
-                ),
-                onPressed: () {
-                  setState(() {
-                    _currentIndex = 2;
-                  });
-                },
-              ),
-              IconButton(
-                icon: Icon(
-                  Icons.person,
-                  color: _currentIndex == 3 ? AppColors.primary : Colors.grey,
-                ),
-                onPressed: () {
-                  setState(() {
-                    _currentIndex = 3;
-                  });
-                  Navigator.push(
-                    context,
-                    MaterialPageRoute(builder: (_) => const ProfileScreen()),
-                  );
-                },
-              ),
-            ],
-          ),
-        ),
+      bottomNavigationBar: BottomNavBar(
+        currentIndex: _currentIndex,
+        onTap: (index) {
+          setState(() {
+            _currentIndex = index;
+          });
+        },
       ),
     );
   }
