+++ pubspec.yaml
@@ -27,13 +27,27 @@ dependencies:
   
   # UI related
   fl_chart: ^0.63.0
-  intl: ^0.18.1
+  intl: any
   flutter_svg: ^2.0.7
   lottie: ^2.6.0
   
   # Utility packages
   uuid: ^3.0.7
   connectivity_plus: ^4.0.2
+
+  #language support
+  easy_localization: ^3.0.3
+
+  # local auth
+  local_auth: ^2.3.0
+  # Add this for Android support
+  local_auth_android: any  
+  # Add this for iOS support
+  local_auth_darwin: any  
+  # getx
+  get: ^4.6.6
+  # flutter_screen_lock
+  flutter_screen_lock: ^9.1.0
 
 dev_dependencies:
   flutter_test:
